backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
pydantic
pytest
